feat: add @veramo/utils package
