+++ packages/did-comm/src/types/utility-types.ts
@@ -1,6 +1,4 @@
-import { IKey, KeyMetadata } from '@veramo/core'
 import { JWE } from 'did-jwt'
-import { VerificationMethod } from 'did-resolver'
 import { DIDCommMessageMediaType, IDIDCommMessage } from './message-types'
 
 /**
@@ -40,33 +38,3 @@ export type _GenericJWS = {
  * @internal
  */
 export type _DIDCommSignedMessage = _FlattenedJWS | _GenericJWS
-
-/**
- * Broader definition of Verification method that includes the legacy publicKeyBase64
- * @internal
- */
-export type _ExtendedVerificationMethod = VerificationMethod & { publicKeyBase64?: string }
-
-/**
- * Represents an {@link @veramo/core#IKey} that has been augmented with its corresponding
- * entry from a DID document.
- * `key.meta.verificationMethod` will contain the {@link did-resolver#VerificationMethod}
- * object from the {@link did-resolver#DIDDocument}
- * @internal
- */
-export interface _ExtendedIKey extends IKey {
-  meta: KeyMetadata & {
-    verificationMethod: _NormalizedVerificationMethod
-  }
-}
-
-/**
- * Represents a {@link did-resolver#VerificationMethod} whose public key material
- * has been converted to `publicKeyHex`
- *
- * @internal
- */
-export type _NormalizedVerificationMethod = Omit<
-  VerificationMethod,
-  'publicKeyBase58' | 'publicKeyBase64' | 'publicKeyJwk'
->
