+++ packages/did-provider-key/tsconfig.json
@@ -3,7 +3,8 @@
   "compilerOptions": {
     "rootDir": "src",
     "outDir": "build",
-    "declarationDir": "build"
+    "declarationDir": "build",
+    "skipLibCheck": true
   },
   "references": [{ "path": "../core" }, { "path": "../did-manager" }]
 }
