+++ packages/utils/README.md
@@ -1,0 +1,3 @@
+# Veramo utils
+
+A package with helper methods.
