+++ packages/utils/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@veramo/utils",
+  "description": "Helper methods for Veramo plugins",
+  "version": "2.1.2",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
+  "scripts": {
+    "build": "tsc"
+  },
+  "dependencies": {
+    "@stablelib/ed25519": "^1.0.2",
+    "@veramo/core": "^2.1.0",
+    "cross-fetch": "^3.1.4",
+    "debug": "^4.1.1",
+    "did-jwt": "5.7.0",
+    "did-resolver": "^3.1.0",
+    "uint8arrays": "^3.0.0",
+    "uuid": "^8.3.0"
+  },
+  "devDependencies": {
+    "@types/debug": "4.1.7",
+    "@types/uuid": "8.3.1",
+    "typescript": "4.4.2"
+  },
+  "files": [
+    "build/**/*",
+    "src/**/*",
+    "README.md",
+    "LICENSE"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": "git@github.com:uport-project/veramo.git",
+  "author": "Mircea Nistor <mircea.nistor@mesh.xyz>",
+  "license": "Apache-2.0",
+  "keywords": []
+}
