+++ packages/utils/src/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Contains helper methods for Veramo plugins.
+ *
+ * @packageDocumentation
+ */
+
+
+export * from './types/utility-types'
+export * from './utils'
